kvutils: Split Deserialization from Serialization.
